Detach MenuLink component and CSS module
